+++ main.py
@@ -1,5 +1,4 @@
 import datetime
-import json
 import os
 import random
 
@@ -10,15 +9,13 @@ import torch.optim.lr_scheduler as lr_scheduler
 
 from ignite.handlers.param_scheduler import create_lr_scheduler_with_warmup
 
-import optim
-from dataset.dataset_loader import dataset_loader, split_data_loader
-from log import log_info_time, time_checker
+from rppg.dataset_loader import dataset_loader, split_data_loader
+from log import time_checker
 from loss import loss_fn
-from models import is_model_support, get_model, summary, get_ver_model
+from models import is_model_support, get_model
 from optim import optimizer
-from utils.dataset_preprocess import preprocessing, dataset_split
-from utils.train import train_fn, test_fn, train_multi_model_fn, test_multi_model_fn, find_lr
-from matplotlib import pyplot as plt
+from rppg.preprocessing.dataset_preprocess import preprocessing, dataset_split
+from rppg.train import train_fn, test_fn, train_multi_model_fn, test_multi_model_fn
 
 from params import params
