+++ model_act.py
@@ -1,5 +1,4 @@
 import datetime
-import datetime
 import json
 import os
 
@@ -14,7 +13,7 @@ from sklearn.model_selection import KFold
 from torch.autograd import Variable
 from torch.utils.data import DataLoader
 
-from dataset.dataset_loader import dataset_loader
+from rppg.dataset_loader import dataset_loader
 from loss import loss_fn
 from models import get_ver_model
 from pytorch_grad.utils.image import show_cam_on_image
